Initial implementation for reusing uuids
The Placeholderservice will recursively travers the tree of data and
replace values when using one of the magic tokens from the existing
yaml-data. If the value is not available it will create a new one.
Currently only uuids are supported, but is extendable.

Also tests are fixed (when using the latest version of phabalicious),
also test will test if newly created data has the same uuids as the
baseline.

Additionally some reorganisation of files were done, to streamline
witht the other tests.